[Feature/72/Sale-System]: Sale system completed

+++ src/main/java/com/daimontech/dsapi/product/model/DiscountPackage.java
@@ -22,5 +22,5 @@ public class DiscountPackage {
     @JoinColumn(name = "packages_id")
     private Packages packages;
 
-    private int discount;
+    private double discount;
 }
